[ci] Send test e-mail to 'vagrant' account
The default 'debops.etc_aliases' configuration redirects all mail for
the local 'root' account to the 'root' account on the DNS domain
instead. In the Vagrant tests, the domain 'vagrant.test' is a false
domain which doesn't exist. Postfix tries to deliver the mails sent to
the 'root' account to that domain and fails, which also fails the test.

This patch changes the destination account to 'vagrant', which is
a local system account without any forwarding configured by default.
This should make the CI test pass.