[debops.roundcube] Fix issue with database types
The new configuration used by the 'debops.roundcube' role had an issue
where using the 'sqlite' database type resulted in undefined dictionary
keys which were present in the 'mysql' database type. This resulted in
an exception during Roundcube configuration generation.

This patch ensures that missing dictionary keys are not fatal, and the
role can work with the 'sqlite' database type.