[debops.lxd] Don't shutdown containers on LXD stop
The 'lxd.service' systemd unit mistakenly stopped the running LXC
containers when the LXD service was restarted. Since this functionality
exists already in the 'lxd-containers.service' unit, the 'ExecStop='
line has been removed from the 'lxd.service' unit. This allows the LXD
daemon to be restarted without changing the state of the existing
containers.