[ferm] Accept DHCPv6 multicast solicit packets
While `accept_dhcpv6_client` rule allows UDP DHCPv6 responses out to
local systems, the initial solicit messages are still dropped on
a default install. The very initial *Solicit* message is sent to
the `ff02::1:2` - a multicast address.

This patch adds a rule in the 'ferm' firewall to accept the initial
DHCPv6 solicit packets.

Ref: https://tools.ietf.org/html/rfc8415#section-7.1