[docs] Reorganize docs for better manpage support
The manpage support in Sphinx is decent, but not ideal. On the other
hand, some parts of the DebOps documentation like the default variable
pages generated from YAML look horrible in the manpage format, therefore
they will be included only in the HTML documentation. This changes the
order of pages a bit, but the documentation should be fine as long as it
is consistent.