[postgresql_server] Support more efficient backukp compression
Add configuration options to configure autopostgresqlbackup to compress
the backups directly with pg_dump instead of first dumping an
uncompressed file and then compressing it.

This uses the autopostgresqlbackup configuration option COMMCOMP which
adds the --compress option to pg_dump. Unfortunately this is labeled as
"communication compression" in the original file, but it has nothing to
do with compressing the communication. See the pg_dump manpage for
details about the --compress option.

(cherry picked from commit c98de30875f5549fda0f2c206497a1b9d49edbe1)