Reorder 'machine' role in 'common.yml' playbook
The 'machine' role defines an Ansible local fact which is used by the
'ldap' role to configure its own local fact which is then used by the
'sshd' role.

This patch changes the order of the 'machine' role in the 'common.yml'
playbook to ensure that it is executed before the 'ldap' role. This
should fix an idempotency issue where the 'ldap' and 'sshd'
configuration was changed on the second run of the 'common.yml'
playbook to apply changes introducted by the 'machine' role.

The 'machine' role is also added to the 'bootstrap-ldap.yml' playbook to
ensure that the desired configuration is applied at the host
provisioning stage.

(cherry picked from commit 19fa6b09c73c1d2bbb58a02df374585263b6401d)