[postfix] Refactor persistent storage code
The 'postfix' role used multiple tasks to check for the existence of the
JSON files that store persistent configuration on the Ansible
Controller. Thanks to a new functionality in the lookup("file") plugin
that can handle non-existent files without stopping Ansible, this can
now be done without any tasks, making the role execution faster.