[postconf] Reorder variables in the playbook
The 'postfix' role is used two times in the 'postconf.yml' playbook
- once to prepare the environment, and the second time to perform the
actual configuration of Postfix service. The problem is that the
dependent variables that define additional configuration managed by the
'postconf' role need to be specified both times in the playbook to be
consistent and idempotent - the persistent configuration stored on
Ansible Controller needs to be merged with the dependent configuration
at the environment stage, and after that the dependent variables need to
be available during Postfix service configuration to install additional
packages and generate persistent configuration files on the Ansible
Controller.

To avoid duplication, this patch moves all of the dependent variables
from various roles to the 'vars:' dictionary defined on the playbook
level. That way, all dependent variables should be visible to all roles
and provide a consistent environment. Since the interaction between the
roles and their variables is strictly managed, there shouldn't be any
conflicts or unexpected outcomes. The users should still be able to
override dependent variables via Ansible inventory.

This is also an experiemnt to see if this style of variable passing does
not bring any issues; if it's successful, all playbooks might be
rearranged in this way in the future.