[CI] Re-enable role checking using 'ansible-lint'
It seems that the issue with 'ansible-lint' not correctly detecting
custom Ansible modules in roles imported using the 'import_role' module
only occurs when the playbooks themselves are tested. This patch changes
the 'ansible-lint' test to only look at the roles which should be
sufficient for now. Check for the playbook code will probably be
implemented separately in the future.