Clarify GPL 3.0 license terms in DebOps codebase
It turns out that due to a quirk in history, most of the codebase
included in DebOps was licensed under GPL 3.0 "only" terms of the
license. A long time ago the first COPYRIGHT file was created with these
terms and included in an Ansible role, and it was just copied over and
over to new roles, and here we are.

This patch clarifies the terms of the GPLv3 license in files that don't
have it included explicitly. The SPDX license identifier is changed from
'GPL-3.0-or-later' to 'GPL-3.0-only' to comply with the implicit license
terms from the "COPYRIGHT" files included in each subdirectory.