Reorder 'netbase' and 'resolvconf' roles in plays
The 'netbase' and 'resolvconf' roles should be executed before the
'secret' role on the common and bootstrap playbooks to set the host name
and domain before the 'secret' role creates directories on the Ansible
Controller. Otherwise the host rename causes the 'pki' role to use wrong
directory names during operation.