[roundcube] Discard cleandb.sh stdout
The 'Roundcube daily database housekeeping' cronjob has output similar
to this:
```
0 records deleted from 'contacts'
0 records deleted from 'contactgroups'
0 records deleted from 'identities'
```
The output is sent by email to the roundcube user, whose mailbox is
usually forwarded to the postmaster or system administrator. The
receiver of these emails may find them annoying.

This change discards the output by redirecting it to /dev/null. Error
output however should still be forwarded.