[pki] Ensure Subject CN falls back to default_domain
The only time 'CN=...' would be anything other than $(hostname -f) was
if the name of the realm was specifying a domain-name (i.e. including a
'.'), or if the subject-variable was including the CN specifically in
the realm configuration. This behavior is however not according to what
the documentation says.

This is said in the documentation;

  subject, acme_subject

    Optional. The Distinguished Name of the certificate, specified as a
    list of DN elements. If not specified, a CommonName based on the
    default domain of the given PKI realm will be used. Empty string
    elements of the list will be ignored.

This patch should now make the documentation true and set the value of
pki_default_domain as CN as a fallback.