[nginx] defaults: Repeat mozilla_intermediate in mozilla_modern.
Using nginx default cipher suite is not suitable for debops
general nature, because one might accidentially activate TLSv1.2
or lower while using mozilla_modern cipher suite. That would
create a security hole.

Copying intermediate to modern is alright, because this list
is specific to TLSv1.2 and lower. TLSv1.3 uses its own ciphers,
which are not supposed to be specified by user.