[dhcrelay] Add debops.dhcrelay role
`debops.dhcrelay` manages the ISC DHCP Relay Agent on Debian Stable for
forwarding of DHCPv4 messages between networks. The init script for
the isc-dhcp-relay package in Buster (and Bullseye, for that matter)
currently does not support DHCPv6, so this role currently only supports
managing dhcrelay for DHCPv4 relaying. A workaround for this can be
implemented in DebOps if there is demand for it.

This role replaces the dhcrelay functionality in `debops.dhcpd`. The
dhcrelay functionality in `debops.dhcpd` will be removed in the next
role update.