[dhcpd] Rewrite large parts of the role
The `debops.dhcpd` role has been largely rewritten in order to modernize
it and to support both DHCPv4 and DHCPv6 servers on the same host. Care
has been taken to improve the documentation and the readability of the
Jinja templates.

Furthermore, the DHCP Relay Agent functionality has been removed. That
functionality can now be found in the `debops.dhcrelay` role. The iPXE
server part is mostly unchanged.