[LDAP] Fix 'nextuidgid.schema' OID number
The schema used the base OID from the 'mailservice.schema' LDAP schema.
By sheer luck that schema uses .3 suffix for attributes and .4 suffix
for objects, so the collision was not detected by OpenLDAP.

This patch changes the 'nextuidgid.schema' base OID to be unique and
adds a corresponding entry in the DebOps OID Registry.