[rsyslog] Fix DNS SRV record detection
This patch fixes the issue in the 'rsyslog' role present when a host
does not have a defined DNS domain (the 'ansible_domain' variable is
empty). In such case the 'rsyslog__syslog_srv_rr' variable would result
in an empty list instead of the 'NXDOMAIN' value which would stop
Ansible with an error "list object has no element 0".