[journald] Add condition to FSS verification task
The FSS verification task should not be executed during normal
operation, but it has been if Ansible was told to run specific tags via
the '--tags' argument. In that case tasks tagged with 'never' tag were
executed as well, and the 'never' tag was ignored.

This patch adds a condition to only run the task if persistent log
storage is enabled, since only then checking the FSS status makes sense.