[fcgiwrap] Fix AnsibleFilterError in check mode
This solves an issue where the 'Generate systemd units' task fails with
this error: `msg: 'AnsibleFilterError: Version comparison: LooseVersion
instance has no attribute ''version'''`

The task tries to template a systemd service file, but the template
needs fcgiwrap__register_version to do a version comparison. This
variable is never initialized in check mode. This commit changes that by
forcing Ansible to run the 'Check the fcgiwrap version' in check mode as
well.