[opendkim] Remove dependency on pyOpenSSL
This dependency was required in Ansible < 2.8.0 because these versions
were unable to use the `cryptography` module, but DebOps is nowadays
developed against Ansible 2.9. pyOpenSSL was used only to generate
private RSA keys for the `debops.opendkim` role. Switching to
`cryptography` is also a security precaution and the Python
Cryptographic Authority [recommends](https://github.com/pyca/cryptography/blob/master/docs/faq.rst#why-use-cryptography)
doing so.

The openssl_privatekey module in `debops.opendkim` will use
`cryptography` when pyOpenSSL is not installed. The Ansible package
depends on this module, so no further changes are required.

Note that this breaks compatibility with the Ansible 2.7 in Debian
Stable, but I believe we aren't trying to maintain compatibility with
that version anymore anyhow.