[lvm] Fix check mode bug in fs creation task
This fixes an issue where the role would fail in check mode. The role
tries to simulate creating a filesystem, but this failed when the
underlying LVM volume did not actually exist (which is to be expected
when running in check mode).

(cherry picked from commit 77ed245eaaa0defb878f61a6da727041f8e65016)