[journald] Create secret directories after facts
The task that creates the directories on the Ansible Controller using
the 'secret' role depends on the Ansible local facts on the remote host.
This patch changes the order of the tasks in the 'journald' role to
ensure that all facts are known and up to date when the 'secret' role is
invoked.