[system_users] fix python version check
When running the system_users role on a controller with only python3
against a host with python2, the wrong getent_passwd.py script
will be exeucted.

The reason is that the "Gather local Ansible user details" task is
delegated to localhost, but the ansible_python_version variable
comes from the host, not the controller.

(cherry picked from commit 6a31829b4145626b649824a39bcfffd53364a670)