[ci] Install Cryptography from Debian in Vagrant
Cryptography is an Ansible dependency and was previously installed as a
Python package. I think we all prefer installing it from the Debian
archive. There's another reason for this since the release of
Cryptography 3.4, because it now requires Rust toolchain 1.45.0+ in
order to build. The Rust toolchain version in Debian 10 is too old for
that.

(cherry picked from commit a594af0b4dca657408b4716c9b6c2cc4d7ed8c2b)