Ensure that UNIX environment is setup for PHP apps
The 'php' role can create the UNIX group, account and home directory for
a given application when properly configured. Due to the recent move to
use handlers instead of tasks for various actions, the 'php' role needs
to prepare the UNIX environment before the service is restarted,
otherwise the playbook runs are interrupted and incomplete.