[rabbitmq_management] Install server locally
After the change to use handlers instead of tasks, I've found a bug in
the 'rabbitmq_management.yml' playbook. The problem comes from use of
a conditional variable to activate certain roles.

The bug happens in the 'nginx' role where for some reason the handler
for refreshing host facts is not applied correctly. Because the 'nginx'
fact provides the role with the package version number, on the
subsequent run the environment is changed and role needs to update the
webserver configuration files.  This causes the subsequent run of the
playbook to be idempotent and breaks CI pipeline.

It seems that the 'flush_handlers' meta action does not like conditions;
removing that condition from the playbook causes for the handler to
execute as expected.

This change results in the 'rabbitmq_management.yml' playbook to install
the RabbitMQ service locally by default. This is most likely a desired
behaviour; if users want to have a completely separate management node
without the service, a separate playbook will be created later to allow
for that.

(cherry picked from commit 1edcfc2e8c25bdbb7e70ff71a98b1e40fbcd6316)