[postldap] Switch back to /<domain>/<user>/ format
The current implementation of Dovecot configuration in the 'dovecot'
role uses the '/%d/%n/Maildir' format as the user's mail storage path.
Since the 'postldap' changes don't affect Dovecot configuration that
much, I think that sticking to the currently used Maildir location will
be best. Keeping the current mailbox layout should avoid the need to
overhaul the mail storage significantly.

This can change in the future DebOps releases, but in that case such
changes should be synchronized between roles.