[ci] Use 'ansible-lint' in autodetection mode
The 'ansible-lint -v' command automatically detects all Ansible content
in the repository, so we don't need to statically specify which
playbooks to test.

The autodetection mode is used by 'ansible-lint' project to test the
DebOps repository, so we should use the same mode to avoid causing
issues for them by mistake.