[ldap] Don't use 'ansible_user' as LDAP admin
In some setups, a shared administrator account can be used to manage
hosts, specified by the 'ansible_user' variable. This might cause issues
when LDAP tasks are meant to be applied for a given host - the shared
account might not have a corresponding LDAP account and even if it did,
it would mean that the password is also shared between multiple
administrators.

To avoid these issues, the 'ldap' role will explicitly not depend on the
'ansible_user' variable and instead will use the local UNIX account as
the username. In a multi-admin environment, each administrator is
expected to have their own LDAP account which can then be used to access
the LDAP directory. If a single LDAP administrator account is needed, it
can still be defined using the '$DEBOPS_LDAP_ADMIN_BINDDN' environment
variable.