[rsyslog] Re-enable streamDriverPermittedPeers
Commit 2cadbbbd23900fa093fe63c2131ab04c922eecc6 added a conditional that
resulted in streamDriverPermittedPeers being set only when the
(undocumented) `drivermode` parameter was set in the `rsyslog__forward`
rule entry. As streamDriverPermittedPeers specifies the name of the
remote peer as used in TLS certificate validation, this option should
always be set when securing the syslog traffic with TLS, which is what
this commit takes care of.

This commit also documents the new rsyslog__forward parameters, and adds
one new default variable: `rsyslog__default_driver_mode`. I have taken
the liberty to rename element.driver, element.drivermode and
element.driverauthmode to better match the naming in the
defaults/main.yml file.

/cc @bleuchtang

Ref: https://github.com/debops/debops/pull/1729