[docs] Fix dead links found by linkchecker
I am as annoyed with hitting 404s as the next guy, so I took some time
to fix some dead links in the documentation. I used linkchecker for
this. The domain names drupal.org, github.com and patreon.com had to be
excluded from link checking due to rate limiting and other 403-like
errors.

I couldn't find alternatives for these URLs:

https://lists.debops.org/pipermail/debops-users/2015-October/000049.html
https://docs.debops.org/en/latest/debops-keyring/docs/entities.html
https://docs.debops.org/en/latest/debops-keyring/docs/index.html
https://api.debops.org/*

I left them as-is.

This also updates the Ansible module targets in
docs/includes/global.rst. The module names are now written in collection
style, e.g. 'some.namespace.module'. The alternative (stripping off the
namespace part) was not really possible as there are modules with the
same name in multiple namespaces.

Ref: https://github.com/debops/docs/pull/290