split-protocols
[dovecot] split imap(s) and pop3(s) into separate protocols

Currently the dovecot_protocols variable bundles the imap/imaps and
pop3/pop3s protocols together. Separating the four protocols makes
it easier to enable/disable a given protocol (e.g. to conform
with RFC8314).

It might look like "dovecot_imap_listeners" and "dovecot_pop3_listeners"
would allow the individual protocols to be enabled/disabled, but that
is not the case (will be addressed in a later patch).