Control checksum computation for ciphertext_block_device
In task _Check if ciphertext block device exists_ the `stat` of the block device computes a checksum by default. When the `ciphertext_block_device` is a large file, e.g. `/tmp/ciphertext_file.raw` of hundreds of GB, this can take prohibitively long time while not adding much to the existence check.

With the proposed fix the checksum computation in `stat` becomes controllable with the variable `check_with_checksum` of items in `cryptsetup__devices`.