[reprepro] Add missing architectures
This adds the expected architectures for Bookworm, and adds the missing
architectures for older releases. The lists of official architectures
were obtained from the release notes. The list of architectures for
Bookworm was obtained from the current Bookworm Release file (which is
subject to change).

Adding the Bookworm architectures is necessary in order to make reprepro
work on Debian 11. This is because Bookworm is configured as the 'next'
release, and if that release doesn't have architecture information the
reprepro commands will fail.

(cherry picked from commit dade0ed9acb6ff768cd354a77280dd45a5713e67)