Switch the CWD to the project directory
Ansible can be executed from anywhere with only the $ANSIBLE_CONFIG
variable that defines the path to its configuration file. However DebOps
and its custom lookup plugins use the Current Working Directory (CWD)
value to detect the project directory. For these two to work correctly,
we need to change the CWD in Python to a sane value.