[sshd] Change the condition for '/run/sshd/' dir
The 'sshd' command requies the '/run/sshd/' directory to exist to not
exit with an error. On systems with already running sshd, this directory
is already created but on specific environments, for example in Docker
containers, it might not be present.

This patch changes the condition when the directory will be created.
Previous condition did not work as expected in Docker, the directory was
not created during the role execution. Changing the condition to check
for 'systemd' as the service manager and create the directory manually
otherwise seems like a reasonable workaround.