[rspamd] Add new role
Rspamd is an advanced spam filtering system that allows evaluation of messages
by a number of rules including regular expressions, statistical analysis and
custom services such as URL black lists. Each message is analysed by Rspamd and
given a spam score.

According to this spam score and the user’s settings, Rspamd recommends an
action for the MTA to apply to the message, for example, to pass, reject or add
a header. Rspamd is designed to process hundreds of messages per second
simultaneously, and provides a number of useful features.

Rspamd also includes a web-based UI.

This role features integration with debops.postfix, debops.redis, debops.nginx,
etc, allowing for a working configuration out-of-the-box.

Partially based on: https://gitlab.com/ciphermail/debops.rspamd

With some manual configuration, DKIM generation/rotation is also supported,
even with automatic DNS updates (after the initial manual setup).