[sshd] Improve LDAP bind password file generation
First, to echo a password means it'll show up temporarily in the process list.

Second, to first create the file and then set the right permissions is also not
a good idea.

This patch also removes a variable which seems kind of pointless. Roles usually
don't make file paths configurable, and I can't see why anyone would like to
change it. If it's better to keep the variable around, I can rework the patch.