[ldap-init] Allow plaintext password storage
This introduces two changes to how ldap/init-directory.yml works:

First, it now supports optionally generating a random password.

Second, optional support is added for storing the random password unencrypted
in the /secret directory.

This makes it easier to e.g. run the Ansible Controller in an automated
fashion.

See the docs updates for more details.