[apt_install] Autodetect presence of virtual RNG
apt_install currently installs "haveged" unconditionally on e.g.
KVM guests. There is a comment in defaults.yml to the effect that
while the host can provide entropy to the guest, this cannot be
assumed.

This patch enables checking for if a virtual RNG is present in
the guest, and skips the installation of "haveged" in that case.

Version 2: uses Ansible facts instead of extra tasks.