[icinga_web] Improve LDAP defaults
The LDAP encryption and port parameters are normally already available
as local facts, so use that as the default.

I noted this issue when trying to setup Icinga in a test environment
where LDAP was only available over LDAPS. /etc/ldap/ldap.conf was
correctly setup, but the Icinga defaults did not adapt.

The problem with incorrect LDAP settings is that the resulting errors
are not very straightforward to diagnose. The "Register Icinga node
in Icinga Director" task from debops.icinga will fail. By increasing
the verbosity, one can see that the Icinga director returns a HTTP 200
response rather than 201 (the expected).

It's only after digging through the Icinga web log files that you
can see that the web app fails to initialise the LDAP connection.

Also, I took the liberty of removing the "recommended" for STARTTLS,
since the general trend is towards recommending implicit TLS (i.e.
ldaps) over explicit TLS, as STARTTLS is subject to downgrade attacks:
https://fy.blackhats.net.au/blog/html/2021/08/12/starttls_in_ldap.html