[grub] Fix password generation
Not sure how this could have worked before, without the "-e" option,
echo won't parse the newlines. With this patch applied I can create
grub users with passwords.

Note that the whole shell thing is still unsafe since it'll leak
the grub password to the process list, a more thorough fix would
be to template the password to a tmp file and then cat that to
grub-mkpasswd-pbkdf2 (I've tested it manually and it works).

That'll have to wait for another PR though.