[apt_install] Remove ranger from package list
Installing ranger by default seems somewhat niche, more like personal
preference than something that'd be generally or universally useful.

In addition, apt_install__shell_packages already include the installation of
"mc", which (according to Debian popcon) is 25x as popular as "ranger", and
which is also a console file manager.