[postconf] Disable EHLO IP address check
Rejecting messages based on the resolved EHLO hostname is prohibited by
https://datatracker.ietf.org/doc/html/rfc5321#section-4.1.4:

An SMTP server MAY verify that the domain name argument in the EHLO
command actually corresponds to the IP address of the client. However,
if the verification fails, the server MUST NOT refuse to accept a
message on that basis.

Some Office 365 (outlook.com) servers have EHLO hostnames that resolve
to 127.0.0.1, which caused our mail servers to reject a few (legitimate)
messages.

This change disables the EHLO IP address check on the next playbook run.