Introducing Miniflux service to DebOps
Miniflux is a minimalist and opinionated feed reader. It is written in
Golang which is distributed as a single binary. This commit provides a
playbook and role to run Miniflux service on Debian and Ubuntu
distributions.

Even though this is the first commit of the playbook/role, it is not
ready for production use.

For more information about Miniflux, please check http://miniflux.app/
website.