[slapd] Reduce log level
The default slapd log level generates quite a lot of log spam (the description
for olcLogLevel says that it "...specifies the level at which debugging
statements and operation statistics should be syslogged" [1]).

As an example, setting up a mail host which uses LDAP and receiving a couple
of hundred emails leads to:

24250

Setting the log level to "none" means that high priority messages will still
be logged, despite the somewhat confusing name (again, see [1]).

This seems like a more reasonable default for a production system.

[1]
https://www.openldap.org/doc/admin24/slapdconf2.html#olcLogLevel:%20%3Clevel%3E