[environment] Do not parse again the `variables` fact
Due to the different formats supported for `environment__variables`,
the `print_environment()` macro is not idempotent, and applying it once
again to the `variables` fact may trigger unwanted behaviors.

In fact, the `variables` fact is supposed to already have the correct
structure: it is a name-value dictionary of variables (such as the ones
output by the `print_environment()` macro), and it can therefore just
be used as is in the templates.