Switch 'from/to_json' to 'from/to_yaml' filters
The new Jinja2 3.x release changed the behaviour of 'to_json' filter
- it doesn't accept YAML lists as input anymore, only YAML dictionaries.
To fix this issue, we need to switch the filters to use YAML as
a transport between various places in the code, which permits lists to
be passed between variables.

There will be more places in the code that this might need to happen,
they will be dealt with on later commits.