[sshd] Fix molly-guard when using dracut
molly-guard is currently installed by default by the sshd role.

On machines using dracut (not default), molly-guard will unfortunately mean
that a reboot is no longer possible, see:
https://bugs.debian.org/992351

This patch ensures that a workaround is in place (noop on systems not
using dracut). Also, I've defined the new variable sshd__combined_packages
to make the resulting tasks/main.yml look a bit nicer.

Not 100% happy about adding three more task steps for this corner-case, but
a non-working reboot is on the other hand a pretty nasty scenario.