Add Mac architecture to docker arch map
The latest Mac's use arm64 and this is not present inside the
'docker_server__upstream_arch_map' which means that the correct
'docker_server__upstream_repository' url cannot be created.

The solution is to simply add an option for 'aarch64' which maps to
'arm64' so the correct url can be generated.