pki: convert intermediate to PEM
The pki-realm script tried to convert the intermediate certificate
by running openssl to see whether it could read the file.
Ubuntu's OpenSSL 3.0.2 15 Mar 2022 (Library: OpenSSL 3.0.2 15 Mar 2022)
will happily read a DER file, though, and thus, the test passes
which causes the conversion to be skipped.

This change explicitly requests openssl to read a PEM format.
If that fails, the conversion will be pursued.
This change also explicitly requests PEM format to be produced.

The problem this solves is that applications may not read the
concatenated certificates correctly which results in untrusted connections.