[slapd] Update docs
The OpenLDAP docs has changed the terminology from
"N-Way Multi-Master" to "N-Way Multi-Provider" (and invalidated
the link anchors in the process).

This updates the debops.slapd docs to be in line with the upstream
docs (to avoid confusion), and also fixes some minor nits in the process.

(cherry picked from commit 6a512c763df58d54260c538dc8da2969660e3fc6)