[apparmor] Remove option to enforce all profiles
Both the official documentation and the role itself warn against
doing this. In addition, all the relevant profiles (i.e. the ones
set to just complain by default) live under /usr/share/apparmor/
and would need to be copied to /etc/apparmor.d/ first.

So, let's just remove this option for now. It's a footgun with
questionable value.