[dovecot] Fix passdb lookup
The syntax used for passdb lookups in Dovecot isn't always entirely intuitive,
the expected format is:

<ldap attribute>=<dovecot field/cfg option>[=<value, in a printf-like style>]

So "uid=user" means that Dovecot's "user" field will be assigned the value
of the LDAP "uid" field for a given user.

Now, looking at pass_attrs in the passdb lookup, there's a couple of issues.

First, "uid=%u" would mean that for a user "bob@example.com", Dovecot should
take the value of the "uid" LDAP attribute and assign it to the Dovecot field
"bob@example.com", which appears unlikely to be the intended meaning.

This can also be seen with "dovecot auth" (note the last line):

    # dovecot auth test bob@example.com
      Password: <some password>
      passdb: bob@example.com auth succeeded
      extra fields:
        user=bob@example.com
        bob@example.com=bob

Second, "mail" is not a valid password database field, and on top of that, the
DebOps mailservice.schema doesn't include the "mailDrop" attribute (though it
defines "mailHomeDirectory" which is used by the userdb lookup).

Third, "userPassword" is not a valid password database field, "password" is
though (although it isn't used when auth_bind is "yes", which is the default
in the Dovecot role).

See:
https://doc.dovecot.org/configuration_manual/authentication/password_database_extra_fields/