Fix ASCII password lookup character sets
Multiple invocations of the password lookup used the term `ascii`
instead of `ascii_letters` in the `chars` parameter. Because of this
only the letters a,c,i,s were added to the character set instead of all
lowercase and uppercase ASCII letters.

As all password lookups did at least also include all digits and the
minimum password length used is 20 characters, this did not result in
weak passwords.