Remove warn argument from command and shell module
The `warn` argument to the command and shell modules was deprecated in
Ansible 2.11 and removed in 2.14. This argument is no longer needed as
no warnings are issued anymore if these modules are used in ways which
could be replaced by an Ansible module. For cases where ansible-lint
reports an error without the argument, appropriate `noqa` comments are
added.

See https://github.com/ansible/ansible/pull/70504.