Revert "chore(php): Refactor condition, avoid negation"
This reverts commit 4a1dc1ce14014f97f65a7c6fc8c4a37fda70711f.

The change doesn't work on fresh host installations. The condition
depends on the presence of the 'ansible_local.php.long_version' fact,
which is not present on a fresh installation, so the condition fails
every time. After that the role tries to use the value from the fact,
but since it doesn't exist, Ansible fails with an 'undefined variable'
error.