[sshd] Ensure that '/run/sshd' directory exists
There is an issue in certain LXC containers where after configuring
sshd, it fails to restart due to the '/run/sshd' directory missing. This
is caused by sshd using systemd socket activation in this environments,
and the directory being removed when the last session is closed.

This is a short them solution before the role is redesigned to take the
new configuration into account.