[dist-upgrade] Make sure dbus is installed
I think that use of 'latest' in this task was been added from the start
to ensure that 'dbus' is actually newest to help with upgrades, but that
might have to be ensured in older Debian releases. At this point, we can
probably safely just ensure the presence of the package with available
version.