[icinga_db,icinga_web] Support remote databases and SSL
With the preparatory patches out of the way, this patch brings the
roles in line with current requirements (ansible-lint, fully-qualified
module names, etc).

The roles now support remote DBs for PostgreSQL/MariaDB, and
automatically uses a SSL connection if supported (if the DB
is on a remote host and the DebOps pki role is enabled).

The x509 module in icinga_web is also ready to support PostgreSQL as
soon as the next version is released by the upstream project.