Add explicit 'changed_when:' statements to tasks
Many roles in DebOps use 'ansible.builtin.shell:' or
'ansible.builtin.command:' tasks to execute commands on hosts. These
tasks are executed by conditions resolved on Ansible level and are not
written to handle idempotency themselves. The 'ansible-lint' linter
recommends that these tasks should still handle idempotency properly and
report that state to Ansible.

This patch adds 'changed_when:' conditions (with 'register:' where
necessary) to evaluate the idempotency state in the above cases. The
current solution is simple and only checks the state of the 'changed:'
return value. The conditions for different tasks will be improved in the
future, but this might require redesign of the logic for specific
situations.